+++ integrations/google-analytics/lib/index.js
@@ -63,7 +63,6 @@ var GA = (exports.Integration = integration('Google Analytics')
   .option('optimize', '')
   .option('nameTracker', false)
   .option('resetCustomDimensionsOnPage', [])
-  .option('sendEventCategoryEcommerce', false)
   .tag('library', '<script src="//www.google-analytics.com/analytics.js">')
   .tag('double click', '<script src="//stats.g.doubleclick.net/dc.js">')
   .tag('http', '<script src="http://www.google-analytics.com/ga.js">')
@@ -693,12 +692,12 @@ GA.prototype.pushEnhancedEcommerce = function(track, opts, trackerName) {
 
   // Send an event level category for events with top level `category` fields per Segment Ecommerce Spec.
   // This field should be generic and not scoped to product level category: https://developers.google.com/analytics/devguides/collection/analyticsjs/enhanced-ecommerce#product-click.
+  var event = track.event().toLowerCase();
   if (
-    opts.sendEventCategoryEcommerce &&
-    (track.event().toLowerCase() === 'product clicked' ||
-      'product added' ||
-      'product viewed' ||
-      'product removed')
+    event === 'product clicked' ||
+    event === 'product added' ||
+    event === 'product viewed' ||
+    event === 'product removed'
   ) {
     args[2] = 'EnhancedEcommerce';
   }