+++ integrations/google-analytics/test/index.test.js
@@ -58,7 +58,6 @@ describe('Google Analytics', function() {
         .option('optimize', '')
         .option('nameTracker', false)
         .option('sampleRate', 100)
-        .option('sendEventCategoryEcommerce', false)
         .option('resetCustomDimensionsOnPage', [])
     );
   });
@@ -1141,64 +1140,6 @@ describe('Google Analytics', function() {
           analytics.deepEqual(toArray(window.ga.args[4]), [
             'send',
             'event',
-            'cat 1',
-            'product added',
-            {
-              dimension1: 'true',
-              metric1: 'true',
-              nonInteraction: 1
-            }
-          ]);
-        });
-
-        it('should set event category to EnhancedEcommerce when setting enabled and event product added', function() {
-          ga.options.sendEventCategoryEcommerce = true;
-          ga.options.setAllMappedProps = false;
-          ga.options.dimensions = { testDimension: 'dimension1' };
-          ga.options.metrics = { testMetric: 'metric1' };
-
-          analytics.track('product added', {
-            currency: 'CAD',
-            quantity: 1,
-            price: 24.75,
-            name: 'my product',
-            category: 'cat 1',
-            sku: 'p-298',
-            testDimension: true,
-            testMetric: true,
-            position: 4
-          });
-
-          analytics.assert(window.ga.args.length === 5);
-          analytics.deepEqual(toArray(window.ga.args[1]), [
-            'set',
-            '&cu',
-            'CAD'
-          ]);
-          analytics.deepEqual(toArray(window.ga.args[2]), [
-            'ec:addProduct',
-            {
-              id: 'p-298',
-              name: 'my product',
-              category: 'cat 1',
-              quantity: 1,
-              price: 24.75,
-              brand: undefined,
-              variant: undefined,
-              currency: 'CAD',
-              metric1: 'true',
-              dimension1: 'true',
-              position: 4
-            }
-          ]);
-          analytics.deepEqual(toArray(window.ga.args[3]), [
-            'ec:setAction',
-            'add',
-            {}
-          ]);
-          analytics.deepEqual(toArray(window.ga.args[4]), [
-            'send',
-            'event',
             'EnhancedEcommerce',
             'product added',
             {
@@ -1256,7 +1197,7 @@ describe('Google Analytics', function() {
           analytics.deepEqual(toArray(window.ga.args[4]), [
             'send',
             'event',
-            'cat 1',
+            'EnhancedEcommerce',
             'product added',
             {
               dimension1: 'true',
@@ -1304,7 +1245,7 @@ describe('Google Analytics', function() {
           analytics.deepEqual(toArray(window.ga.args[4]), [
             'send',
             'event',
-            'cat 1',
+            'EnhancedEcommerce',
             'product added',
             'sample label',
             { nonInteraction: 1 }
@@ -1312,62 +1253,6 @@ describe('Google Analytics', function() {
         });
 
         it('should send product removed data', function() {
-          ga.options.setAllMappedProps = false;
-          ga.options.dimensions = { testDimension: 'dimension1' };
-          ga.options.metrics = { testMetric: 'metric1' };
-
-          analytics.track('product removed', {
-            currency: 'CAD',
-            quantity: 1,
-            price: 24.75,
-            name: 'my product',
-            category: 'cat 1',
-            sku: 'p-298',
-            testDimension: true,
-            testMetric: true
-          });
-
-          analytics.assert(window.ga.args.length === 5);
-          analytics.deepEqual(toArray(window.ga.args[1]), [
-            'set',
-            '&cu',
-            'CAD'
-          ]);
-          analytics.deepEqual(toArray(window.ga.args[2]), [
-            'ec:addProduct',
-            {
-              id: 'p-298',
-              name: 'my product',
-              category: 'cat 1',
-              quantity: 1,
-              price: 24.75,
-              brand: undefined,
-              variant: undefined,
-              currency: 'CAD',
-              metric1: 'true',
-              dimension1: 'true'
-            }
-          ]);
-          analytics.deepEqual(toArray(window.ga.args[3]), [
-            'ec:setAction',
-            'remove',
-            {}
-          ]);
-          analytics.deepEqual(toArray(window.ga.args[4]), [
-            'send',
-            'event',
-            'cat 1',
-            'product removed',
-            {
-              dimension1: 'true',
-              metric1: 'true',
-              nonInteraction: 1
-            }
-          ]);
-        });
-
-        it('should set event category to EnhancedEcommerce when setting enabled and event product removed', function() {
-          ga.options.sendEventCategoryEcommerce = true;
           ga.options.setAllMappedProps = false;
           ga.options.dimensions = { testDimension: 'dimension1' };
           ga.options.metrics = { testMetric: 'metric1' };
@@ -1423,64 +1308,6 @@ describe('Google Analytics', function() {
         });
 
         it('should send product viewed data', function() {
-          ga.options.setAllMappedProps = false;
-          ga.options.dimensions = { testDimension: 'dimension1' };
-          ga.options.metrics = { testMetric: 'metric1' };
-
-          analytics.track('product viewed', {
-            currency: 'CAD',
-            quantity: 1,
-            price: 24.75,
-            name: 'my product',
-            category: 'cat 1',
-            sku: 'p-298',
-            list: 'Apparel Gallery',
-            testDimension: true,
-            testMetric: true
-          });
-
-          analytics.assert(window.ga.args.length === 5);
-          analytics.deepEqual(toArray(window.ga.args[1]), [
-            'set',
-            '&cu',
-            'CAD'
-          ]);
-          analytics.deepEqual(toArray(window.ga.args[2]), [
-            'ec:addProduct',
-            {
-              id: 'p-298',
-              name: 'my product',
-              category: 'cat 1',
-              quantity: 1,
-              price: 24.75,
-              brand: undefined,
-              variant: undefined,
-              currency: 'CAD',
-              metric1: 'true',
-              dimension1: 'true'
-            }
-          ]);
-          analytics.deepEqual(toArray(window.ga.args[3]), [
-            'ec:setAction',
-            'detail',
-            { list: 'Apparel Gallery' }
-          ]);
-          analytics.deepEqual(toArray(window.ga.args[4]), [
-            'send',
-            'event',
-            'cat 1',
-            'product viewed',
-            {
-              dimension1: 'true',
-              metric1: 'true',
-              nonInteraction: 1
-            }
-          ]);
-          analytics.assert(window.ga.args[1][0] === 'set');
-        });
-
-        it('should set event category to EnhancedEcommerce when setting enabled and event product viewed', function() {
-          ga.options.sendEventCategoryEcommerce = true;
           ga.options.setAllMappedProps = false;
           ga.options.dimensions = { testDimension: 'dimension1' };
           ga.options.metrics = { testMetric: 'metric1' };
@@ -1876,63 +1703,6 @@ describe('Google Analytics', function() {
 
         it('should send product clicked data', function() {
           ga.options.setAllMappedProps = false;
-          ga.options.dimensions = { testDimension: 'dimension1' };
-          ga.options.metrics = { testMetric: 'metric1' };
-
-          analytics.track('product clicked', {
-            currency: 'CAD',
-            quantity: 1,
-            price: 24.75,
-            name: 'my product',
-            category: 'cat 1',
-            sku: 'p-298',
-            list: 'search results',
-            testDimension: true,
-            testMetric: true
-          });
-
-          analytics.assert(window.ga.args.length === 5);
-          analytics.deepEqual(toArray(window.ga.args[1]), [
-            'set',
-            '&cu',
-            'CAD'
-          ]);
-          analytics.deepEqual(toArray(window.ga.args[2]), [
-            'ec:addProduct',
-            {
-              id: 'p-298',
-              name: 'my product',
-              category: 'cat 1',
-              quantity: 1,
-              price: 24.75,
-              brand: undefined,
-              variant: undefined,
-              currency: 'CAD',
-              metric1: 'true',
-              dimension1: 'true'
-            }
-          ]);
-          analytics.deepEqual(toArray(window.ga.args[3]), [
-            'ec:setAction',
-            'click',
-            { list: 'search results' }
-          ]);
-          analytics.deepEqual(toArray(window.ga.args[4]), [
-            'send',
-            'event',
-            'cat 1',
-            'product clicked',
-            {
-              dimension1: 'true',
-              metric1: 'true',
-              nonInteraction: 1
-            }
-          ]);
-        });
-
-        it('should set event category to EnhancedEcommerce when setting enabled and event product clicked', function() {
-          ga.options.setAllMappedProps = false;
-          ga.options.sendEventCategoryEcommerce = true;
           ga.options.dimensions = { testDimension: 'dimension1' };
           ga.options.metrics = { testMetric: 'metric1' };